# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Budget Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReminderPriority =
  (typeof ReminderPriority)[keyof typeof ReminderPriority];

export const ReminderPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Reminder {
  id: number;
  title: string;
  description: string;
  dueDate?: string | null;
  isCompleted: boolean;
  priority: ReminderPriority;
  createdAt: string;
}

export interface Expense {
  id: number;
  amount: number;
  category: string;
  description: string;
  date: string;
  isRecurring: boolean;
  createdAt: string;
}

export interface DashboardSummary {
  totalIncome: number;
  totalExpenses: number;
  totalCharity: number;
  totalDebts: number;
  totalSavings: number;
  balance: number;
  monthlyBudget: number;
  upcomingReminders: Reminder[];
  recentExpenses: Expense[];
}

export interface CreateExpenseInput {
  amount: number;
  category: string;
  description: string;
  date: string;
  isRecurring: boolean;
}

export interface Income {
  id: number;
  amount: number;
  source: string;
  description: string;
  date: string;
  isRecurring: boolean;
  createdAt: string;
}

export interface CreateIncomeInput {
  amount: number;
  source: string;
  description: string;
  date: string;
  isRecurring: boolean;
}

export interface CharityEntry {
  id: number;
  amount: number;
  recipient: string;
  description: string;
  date: string;
  isTithe: boolean;
  createdAt: string;
}

export interface CreateCharityEntryInput {
  amount: number;
  recipient: string;
  description: string;
  date: string;
  isTithe: boolean;
}

export type DebtType = (typeof DebtType)[keyof typeof DebtType];

export const DebtType = {
  owed_to_me: "owed_to_me",
  i_owe: "i_owe",
} as const;

export type DebtStatus = (typeof DebtStatus)[keyof typeof DebtStatus];

export const DebtStatus = {
  active: "active",
  paid: "paid",
} as const;

export interface Debt {
  id: number;
  name: string;
  type: DebtType;
  totalAmount: number;
  remainingAmount: number;
  dueDate?: string | null;
  status: DebtStatus;
  notes: string;
  createdAt: string;
}

export type CreateDebtInputType =
  (typeof CreateDebtInputType)[keyof typeof CreateDebtInputType];

export const CreateDebtInputType = {
  owed_to_me: "owed_to_me",
  i_owe: "i_owe",
} as const;

export type CreateDebtInputStatus =
  (typeof CreateDebtInputStatus)[keyof typeof CreateDebtInputStatus];

export const CreateDebtInputStatus = {
  active: "active",
  paid: "paid",
} as const;

export interface CreateDebtInput {
  name: string;
  type: CreateDebtInputType;
  totalAmount: number;
  remainingAmount: number;
  dueDate?: string | null;
  status: CreateDebtInputStatus;
  notes: string;
}

export type SavingType = (typeof SavingType)[keyof typeof SavingType];

export const SavingType = {
  savings: "savings",
  asset: "asset",
  investment: "investment",
} as const;

export interface Saving {
  id: number;
  name: string;
  category: string;
  type: SavingType;
  targetAmount?: number | null;
  currentAmount: number;
  notes: string;
  createdAt: string;
}

export type CreateSavingInputType =
  (typeof CreateSavingInputType)[keyof typeof CreateSavingInputType];

export const CreateSavingInputType = {
  savings: "savings",
  asset: "asset",
  investment: "investment",
} as const;

export interface CreateSavingInput {
  name: string;
  category: string;
  type: CreateSavingInputType;
  targetAmount?: number | null;
  currentAmount: number;
  notes: string;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  color: string;
  isPinned: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteInput {
  title: string;
  content: string;
  color: string;
  isPinned: boolean;
}

export type CreateReminderInputPriority =
  (typeof CreateReminderInputPriority)[keyof typeof CreateReminderInputPriority];

export const CreateReminderInputPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateReminderInput {
  title: string;
  description: string;
  dueDate?: string | null;
  isCompleted: boolean;
  priority: CreateReminderInputPriority;
}

export interface Settings {
  id: number;
  currency: string;
  monthlyBudget: number;
  incomeForTithe: number;
  tithePercentage: number;
  userName: string;
}

export interface UpdateSettingsInput {
  currency: string;
  monthlyBudget: number;
  incomeForTithe: number;
  tithePercentage: number;
  userName: string;
}

export type GetExpensesParams = {
  month?: string;
  category?: string;
};

export type GetIncomesParams = {
  month?: string;
};

export type GetCharityEntriesParams = {
  month?: string;
};