# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Budget Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: Get all expenses
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseInput"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Incomes
  /incomes:
    get:
      operationId: getIncomes
      tags: [incomes]
      summary: Get all incomes
      parameters:
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of incomes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Income"
    post:
      operationId: createIncome
      tags: [incomes]
      summary: Create income
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeInput"
      responses:
        "201":
          description: Created income
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
  /incomes/{id}:
    put:
      operationId: updateIncome
      tags: [incomes]
      summary: Update income
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeInput"
      responses:
        "200":
          description: Updated income
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
    delete:
      operationId: deleteIncome
      tags: [incomes]
      summary: Delete income
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Charity
  /charity:
    get:
      operationId: getCharityEntries
      tags: [charity]
      summary: Get all charity entries
      parameters:
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of charity entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CharityEntry"
    post:
      operationId: createCharityEntry
      tags: [charity]
      summary: Create charity entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharityEntryInput"
      responses:
        "201":
          description: Created charity entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharityEntry"
  /charity/{id}:
    put:
      operationId: updateCharityEntry
      tags: [charity]
      summary: Update charity entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharityEntryInput"
      responses:
        "200":
          description: Updated charity entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharityEntry"
    delete:
      operationId: deleteCharityEntry
      tags: [charity]
      summary: Delete charity entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Debts
  /debts:
    get:
      operationId: getDebts
      tags: [debts]
      summary: Get all debts
      responses:
        "200":
          description: List of debts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debt"
    post:
      operationId: createDebt
      tags: [debts]
      summary: Create debt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDebtInput"
      responses:
        "201":
          description: Created debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /debts/{id}:
    put:
      operationId: updateDebt
      tags: [debts]
      summary: Update debt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDebtInput"
      responses:
        "200":
          description: Updated debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
    delete:
      operationId: deleteDebt
      tags: [debts]
      summary: Delete debt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Savings
...
  /notes:
    get:
      operationId: getNotes
      tags: [notes]
      summary: Get all notes
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    put:
      operationId: updateNote
      tags: [notes]
      summary: Update note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteInput"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Reminders
...
components:
  schemas:
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        date:
          type: string
        isRecurring:
          type: boolean
        createdAt:
          type: string
      required: [id, amount, category, description, date, isRecurring, createdAt]
    CreateExpenseInput:
      type: object
      properties:
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        date:
          type: string
        isRecurring:
          type: boolean
      required: [amount, category, description, date, isRecurring]
    Income:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        source:
          type: string
        description:
          type: string
        date:
          type: string
        isRecurring:
          type: boolean
        createdAt:
          type: string
      required: [id, amount, source, description, date, isRecurring, createdAt]
    CreateIncomeInput:
      type: object
      properties:
        amount:
          type: number
        source:
          type: string
        description:
          type: string
        date:
          type: string
        isRecurring:
          type: boolean
      required: [amount, source, description, date, isRecurring]
    CharityEntry:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        recipient:
          type: string
        description:
          type: string
        date:
          type: string
        isTithe:
          type: boolean
        createdAt:
          type: string
      required: [id, amount, recipient, description, date, isTithe, createdAt]
    CreateCharityEntryInput:
      type: object
      properties:
        amount:
          type: number
        recipient:
          type: string
        description:
          type: string
        date:
          type: string
        isTithe:
          type: boolean
      required: [amount, recipient, description, date, isTithe]
    Debt:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [owed_to_me, i_owe]
        totalAmount:
          type: number
        remainingAmount:
          type: number
        dueDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, paid]
        notes:
          type: string
        createdAt:
          type: string
      required: [id, name, type, totalAmount, remainingAmount, status, notes, createdAt]
    CreateDebtInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [owed_to_me, i_owe]
        totalAmount:
          type: number
        remainingAmount:
          type: number
        dueDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, paid]
        notes:
          type: string
      required: [name, type, totalAmount, remainingAmount, status, notes]
...
    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        color:
          type: string
        isPinned:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, content, color, isPinned, createdAt, updatedAt]
    CreateNoteInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        color:
          type: string
        isPinned:
          type: boolean
      required: [title, content, color, isPinned]
...