# lib/api-zod/src/generated/types/charityEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Budget Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CharityEntry {
  id: number;
  amount: number;
  recipient: string;
  description: string;
  date: string;
  isTithe: boolean;
  createdAt: string;
}