# lib/api-zod/src/generated/types/createExpenseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Budget Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExpenseInput {
  amount: number;
  category: string;
  description: string;
  date: string;
  isRecurring: boolean;
}