# lib/api-zod/src/generated/types/createIncomeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Budget Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateIncomeInput {
  amount: number;
  source: string;
  description: string;
  date: string;
  isRecurring: boolean;
}